This Function will give the delta Cq values. Always: offtarget Cq - target Cq

helper function.

The delta Cq values can be calculated by randomly assign two values, build all combinations of deta Cq. Directly use the values from the same row or just use the mean values.

delta.Cq.data(
  CqType = NA,
  method = "combinatorial",
  onlyNumeric = FALSE,
  return = FALSE
)

Arguments

CqType

this is the Cq value columns from the input.cq that should be used.

method

can be either "combinatorial", "random", "direct" or "mean" (can be abbriviated). This will impact how the differences are calculated.

onlyNumeric

Will only use samples that are a numerical (for percentages the "%" will be stripped)

return

standard=FALSE will write in global scope! Otherwise will return the dataframe.

Value

standard: returns nothing. Creates a dataframe or creates a a dataframe data.cq.sum in global scope